Begin generating GraphQL from definitions
- Split src/workspace/definitions into module directory
- Define helper to generate typeDefs for a given definition
- Define helper to generate resolvers for a given definition
- Begin with sources definition
