+++ packages/db/package.json
@@ -41,6 +41,7 @@
     "@types/graphql": "^14.0.4",
     "@types/jest": "^23.3.11",
     "@types/node": "^10.12.18",
+    "@types/pluralize": "^0.0.29",
     "@types/pouchdb": "^6.3.2",
     "@types/pouchdb-adapter-leveldb": "^6.1.3",
     "@types/web3": "1.0.20",
@@ -67,6 +68,7 @@
     "leveldown": "^5.2.0",
     "module-alias": "^2.1.0",
     "pascal-case": "^2.0.1",
+    "pluralize": "^8.0.0",
     "pouchdb": "7.1.1",
     "pouchdb-adapter-memory": "^7.1.1",
     "pouchdb-adapter-node-websql": "^7.0.0",
